Moved useful methods to separate utils file and updated tests accordingly.
